Remove cobra user, since it's an unnecessary complexity

+++ airootfs/root/customize_airootfs.sh
@@ -14,16 +14,22 @@ usermod -s /usr/bin/zsh root
 cp -aT /etc/skel/ /root/
 chmod 700 /root
 
-useradd -mU -s /usr/bin/zsh student
-useradd -mUG wheel -s /usr/bin/zsh cobra
+useradd -mUG wheel -s /usr/bin/zsh student
 
-chpasswd < /root/passwords
-rm /root/passwords
+if [ -f "/root/passwords" ]; then
+    chpasswd < /root/passwords
+    rm /root/passwords
+else
+    echo -e "\n##### WARNING #####\n/root/passwords was not provided!\n No root access will be possible!\n"
+fi
 
 # Allow the wheel group to use sudo
-sed -i 's/# \(%wheel ALL=(ALL) NOPASSWD: ALL\)/\1/' /etc/sudoers
+sed -i 's/# \(%wheel ALL=(ALL) ALL\)/\1/' /etc/sudoers
 
+# Enable all pacman servers
 sed -i "s/#Server/Server/g" /etc/pacman.d/mirrorlist
+
+# Set journald storage to volatile
 sed -i 's/#\(Storage=\)auto/\1volatile/' /etc/systemd/journald.conf
 
 # Disable power keys
@@ -31,7 +37,7 @@ sed -i 's/#\(HandleSuspendKey=\)suspend/\1ignore/' /etc/systemd/logind.conf
 sed -i 's/#\(HandleHibernateKey=\)hibernate/\1ignore/' /etc/systemd/logind.conf
 sed -i 's/#\(HandleLidSwitch=\)suspend/\1ignore/' /etc/systemd/logind.conf
 
-systemctl enable pacman-init.service choose-mirror.service nodm.service hostname.service NetworkManager.service
 systemctl set-default multi-user.target
+systemctl enable pacman-init.service choose-mirror.service nodm.service hostname.service NetworkManager.service
 
 echo -e "\n\n#####################################\n#     CUSTOMIZATION SCRIPT DONE     #\n#####################################\n\n"
